feat(news): enhance news management UI and image handling
Improved the user interface for creating, editing, and displaying news
articles by integrating card components for better layout and design.
Added image preview functionality during news creation and editing,
allowing users to see their selected images before submission.

- Updated Create and Edit pages to use Card components for layout
- Implemented image preview feature for better user experience
- Enhanced Show page with improved layout and visibility indicators
- Refactored image handling in NewsController to check for file existence

These changes significantly enhance the usability and aesthetics of the
news management features within the dashboard.

+++ app/Http/Controllers/Dashboard/NewsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Models\News;
+use File;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
@@ -97,7 +98,10 @@ class NewsController extends Controller
         if ($request->hasFile('image')) {
             // Delete old image if exists
             if ($news->image) {
-                unlink(public_path('') . $news->image);
+                $isExist = File::exists(public_path('') . $news->image);
+                if ($isExist) {
+                    unlink(public_path('') . $news->image);
+                }
             }
             $fileName = time() . '.' . $request->file('image')->extension();
             $request->file('image')->move(public_path('images'), $fileName);
@@ -116,9 +120,11 @@ class NewsController extends Controller
     public function destroy(News $news)
     {
         // dd('tes', Storage::disk('public') . '.' . $news->image);
-        dd(Storage::exists('public' . '.' . $news->image), $news->image);
         if ($news->image) {
-            unlink(public_path('') . $news->image);
+            $isExist = File::exists(public_path('') . $news->image);
+            if ($isExist) {
+                unlink(public_path('') . $news->image);
+            }
         }
 
         $news->delete();
